addedPayment

+++ client/src/store.js
@@ -2,15 +2,19 @@ import Vue from 'vue'
 import Vuex from 'vuex'
 import axios from  'axios'
 import router from './router';
+// import Swal from 'vue-sweetalert2'
+import Swal from 'sweetalert2'
 const baseUrlShoes = `http://localhost:3000/api/shoes/`
 const baseUrlCart = `http://localhost:3000/api/carts/`
 Vue.use(Vuex)
+// Vue.use(Swal)
 
 export default new Vuex.Store({
   state: {
     isLogin: false,
     role: '',
-    carts: []
+    carts: [],
+    totalPrice: 0
   },
   mutations: {
     SET_LOGIN(state, payload) {
@@ -21,6 +25,9 @@ export default new Vuex.Store({
     },
     ADD_TO_CART(state,payload) {
       state.carts = payload
+    },
+    GET_TOTAL_PRICE(state,payload) {
+      state.totalPrice = payload
     }
   },
   actions: {
@@ -31,11 +38,23 @@ export default new Vuex.Store({
 
     },
     deleteShoe(context,payload) {
-      axios.delete(`${baseUrlShoes}/${payload}`,{
-        headers:{
-          'token': localStorage.getItem('token')
-        }
+
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        type: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
       })
+      .then((result) => {
+        if (result.value) {
+          axios.delete(`${baseUrlShoes}/${payload}`,{
+            headers:{
+              'token': localStorage.getItem('token')
+            }
+          })
         .then((data) => {
           console.log('deleted')
           router.push('/shoes')
@@ -43,6 +62,8 @@ export default new Vuex.Store({
         .catch((error) => {
           console.log(error)
         })
+        }
+      })
     },
     updateShoe(context,payload) {
 
@@ -59,7 +80,13 @@ export default new Vuex.Store({
         quantity: 1
       })
         .then((cartCreated) => {
-          router.push('/cart')
+          Swal.fire({
+            type: 'success',
+            text: 'Product added to cart',
+            showConfirmButton: false,
+            timer: 1500
+          })
+          // router.push('/cart')
           console.log(cartCreated)
         })
         .catch((error) => {
@@ -76,6 +103,21 @@ export default new Vuex.Store({
           console.log(error)
         })
     },
+    getTotal(context,payload) {
+      let price = 0
+      let userId = localStorage.getItem('id')
+      axios.get(`${baseUrlCart}/${userId}`)
+        .then(({data}) =>{
+          data.forEach((cart) => {
+            price += cart.quantity * cart.productId.price
+          })
+          context.commit('GET_TOTAL_PRICE',price)
+          // console.log(price);
+        })
+        .catch(error=> {
+          console.log(error)
+        })
+    }
